refactor: Move CommandParser to 'common' subproject, decouple CompileCommand from CommandParser.
