+++ common/src/main/java/com/synopsys/integration/common/util/parse/CommandParser.java
@@ -5,19 +5,18 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.detectable.detectable.parser;
+package com.synopsys.integration.common.util.parse;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.text.StringTokenizer;
 import org.apache.commons.text.matcher.StringMatcherFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.synopsys.integration.detectable.detectables.clang.compilecommand.CompileCommand;
 import com.synopsys.integration.util.Stringable;
 
 public class CommandParser {
@@ -32,21 +31,15 @@ public class CommandParser {
     private static final String ENCODED_SPACE_CHAR = "%20";
     private static final String ENCODED_TAB_CHAR = "%09";
 
-    public List<String> parseCommand(CompileCommand compileCommand, Map<String, String> optionOverrides) {
-
-        String commandString = compileCommand.command;
-        if (StringUtils.isBlank(commandString)) {
-            commandString = String.join(" ", compileCommand.arguments);
-        }
-        List<String> commandList = parseCommandString(commandString, optionOverrides);
-        return commandList;
+    public List<String> parseCommandString(String commandString) {
+        return parseCommandString(commandString, new HashMap<>());
     }
 
     public List<String> parseCommandString(String commandString, Map<String, String> optionOverrides) {
-        logger.trace(String.format("origCompileCommand         : %s", commandString));
-        String quotesRemovedCompileCommand = encodeQuotedWhitespace(commandString);
-        logger.trace(String.format("quotesRemovedCompileCommand: %s", quotesRemovedCompileCommand));
-        StringTokenizer tokenizer = new StringTokenizer(quotesRemovedCompileCommand);
+        logger.trace(String.format("origCommand         : %s", commandString));
+        String quotesRemovedCommand = encodeQuotedWhitespace(commandString);
+        logger.trace(String.format("quotesRemovedCommand: %s", quotesRemovedCommand));
+        StringTokenizer tokenizer = new StringTokenizer(quotesRemovedCommand);
         tokenizer.setQuoteMatcher(StringMatcherFactory.INSTANCE.quoteMatcher());
         List<String> commandList = new ArrayList<>();
         String lastPart = "";
