+++ detectable/src/test/java/com/synopsys/integration/detectable/detectables/clang/unit/CommandParserTest.java
@@ -31,9 +31,10 @@ import java.util.Map;
 
 import org.junit.jupiter.api.Test;
 
+import com.synopsys.integration.common.util.parse.CommandParser;
 import com.synopsys.integration.detectable.annotations.UnitTest;
 import com.synopsys.integration.detectable.detectables.clang.compilecommand.CompileCommand;
-import com.synopsys.integration.detectable.detectable.parser.CommandParser;
+import com.synopsys.integration.detectable.detectables.clang.compilecommand.CompileCommandParser;
 
 @UnitTest
 public class CommandParserTest {
@@ -42,7 +43,7 @@ public class CommandParserTest {
         CompileCommand sampleCommand = new CompileCommand();
         sampleCommand.command = "g++ -DDOUBLEQUOTED=\"A value for the compiler\" -DSINGLEQUOTED='Another value for the compiler' file.c -o file.o";
 
-        CommandParser commandParser = new CommandParser();
+        CompileCommandParser commandParser = new CompileCommandParser(new CommandParser());
         List<String> compilerCommand = commandParser.parseCommand(sampleCommand, Collections.emptyMap());
 
         assertEquals("g++", compilerCommand.get(0));
@@ -56,7 +57,7 @@ public class CommandParserTest {
         Map<String, String> optionOverrides = new HashMap<>();
         optionOverrides.put("-o", "/dev/null");
 
-        CommandParser commandParser = new CommandParser();
+        CompileCommandParser commandParser = new CompileCommandParser(new CommandParser());
         List<String> result = commandParser.parseCommand(sampleCommand, optionOverrides);
 
         for (String part : result) {
@@ -78,7 +79,7 @@ public class CommandParserTest {
         CompileCommand command = new CompileCommand();
         command.command = "/usr/bin/clang++-3.6 -DCMAKE_BUILD_TYPE=\\\"Debug\\\" -DCMAKE_CC_FLAGS=\"\\\" -ggdb -Wstrict-aliasing=2 -pedantic -fPIC --std=c11\\\"\"  -c ./pb.cc";
 
-        CommandParser commandParser = new CommandParser();
+        CompileCommandParser commandParser = new CompileCommandParser(new CommandParser());
         List<String> result = commandParser.parseCommand(command, Collections.emptyMap());
 
         assertEquals(5, result.size());
@@ -95,7 +96,7 @@ public class CommandParserTest {
         CompileCommand command = new CompileCommand();
         command.command = "X=\"\\\" a  b\\\"\"";
 
-        CommandParser commandParser = new CommandParser();
+        CompileCommandParser commandParser = new CompileCommandParser(new CommandParser());
         List<String> result = commandParser.parseCommand(command, Collections.emptyMap());
 
         assertEquals(1, result.size());
@@ -110,7 +111,7 @@ public class CommandParserTest {
         command.command = "X=\\\"'a' 'b'\\\"";
         command.file = "test.cc";
 
-        CommandParser commandParser = new CommandParser();
+        CompileCommandParser commandParser = new CompileCommandParser(new CommandParser());
         List<String> result = commandParser.parseCommand(command, Collections.emptyMap());
         assertEquals(1, result.size());
         assertEquals("X=\\\"'a' 'b'\\\"", result.get(0));
